[CHORE] config별 패키지 구분 추가
