--- src/main/java/org/appjam/bongbaek/global/config/SecurityConfig.java
@@ -1,45 +1,0 @@
-package org.appjam.bongbaek.global.config;
-
-import lombok.RequiredArgsConstructor;
-import org.appjam.bongbaek.global.jwt.exception.CustomAccessDeniedHandler;
-import org.appjam.bongbaek.global.jwt.exception.CustomJwtAuthenticationEntryPoint;
-import org.appjam.bongbaek.global.jwt.filter.JwtAuthenticationFilter;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.RequestCacheConfigurer;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-@EnableWebSecurity
-@RequiredArgsConstructor
-@Configuration
-public class SecurityConfig {
-
-    private final JwtAuthenticationFilter jwtAuthenticationFilter;
-    private final CustomJwtAuthenticationEntryPoint customJwtAuthenticationEntryPoint;
-    private final CustomAccessDeniedHandler customAccessDeniedHandler;
-
-    @Bean
-    SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        http.csrf(AbstractHttpConfigurer::disable)
-                .formLogin(AbstractHttpConfigurer::disable)
-                .requestCache(RequestCacheConfigurer::disable)
-                .httpBasic(AbstractHttpConfigurer::disable)
-                .exceptionHandling(exception ->
-                {
-                    exception.authenticationEntryPoint(customJwtAuthenticationEntryPoint);
-                    exception.accessDeniedHandler(customAccessDeniedHandler);
-                });
-
-        http.authorizeHttpRequests(auth -> {
-                    auth.requestMatchers(AuthWhiteList.WHITE_LIST).permitAll();
-                    auth.anyRequest().authenticated();
-                })
-                .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter.class);
-
-        return http.build();
-    }
-}
