+++ src/main/java/org/appjam/bongbaek/global/config/data/QuerydslConfig.java
@@ -1,4 +1,4 @@
-package org.appjam.bongbaek.global.config;
+package org.appjam.bongbaek.global.config.data;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
