+++ src/main/java/org/appjam/bongbaek/global/config/docs/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package org.appjam.bongbaek.global.config;
+package org.appjam.bongbaek.global.config.docs;
 
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.annotations.info.Info;
